requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.5
openai==1.102.0
google-genai==1.32.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
pydantic==2.9.2
